created routes folder and moved endpoints into routes/bars.js
